Added batch book creation endpoint: - BooksController.cs: Feature - allows adding multiple books at once
Updated BookModel and AuthorModel for null safety:
- Models/BookModel.cs, AuthorModel.cs: Prevents null reference errors and solves nullable warnings

Added price validation in NewBookModel:
- NewBookModel.cs: Ensures price is non-negative

Changed DiscountPercentage type to int:
- BookModel.cs: Type consistency

Improved JWT secret validation and usage:
- Program.cs, AccountRepository.cs: Prevents misconfiguration and runtime errors

Refactored repository interfaces and implementations:
- Rpositories/IBooksRepository.cs, Rpositories/BooksRepository.cs,
  Rpositories/IAccountRepository.cs, AccountRepository.cs:
  Better error handling, API consistency, solves nullable warnings

Added EF Core migrations for book table and author relation:
- Migrations/20250806184702_UpdateBookTableToMatchModel.cs,
  Migrations/20250806184702_UpdateBookTableToMatchModel.Designer.cs,
  Migrations/20250820143728_AddAuthorToBook.cs,
  Migrations/20250820143728_AddAuthorToBook.Designer.cs,
  BookstoreContextModelSnapshot.cs:
  Supports new book fields and author relationship

+++ BookstoreServerApiDotnetCore/Models/BookModel.cs
@@ -7,7 +7,7 @@
         public required AuthorModel Author { get; set; }
         public required string Details { get; set; }
         public required decimal Price { get; set; }
-        public decimal DiscountPercentage { get; set; }
+        public int DiscountPercentage { get; set; }
         public required string ImageUrl { get; set; }
     }
 }
